Prospers Implementation, Combining already done!
I tried to understand what we have so far in order not to repeat what has already been done.

In addition to what has already been done, I tried to comment (unprofessionally done) the code to make it easy for anyone to follow. Not exhaustive. The uncommented code being the implementation of the range ... hasnt been well commented due to my insufficient understanding about it.